reader: implement GVariant deserializers
This implements the signature parsers and GVariant deserializers. Some
basic tests for simple types and static compound types are added.
However, proper testing hasn't been done, yet, and is still to be
written.

Expect this to still have some implementation errors. Those will have to
be figured out in followups. Furthermore, once the serializers are done,
we might have to change some state-tracking done in CVariant and
CVariantLevel types.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>