writer: silence gcc warning
Always initialize 'wz' when closing up a container. Apparently gcc seems
to be too stupid to figure out "prev->enclosing" cannot change. All
functions are provided inline, and LLVM+clang is smart enough to figure
it out.. eh.. doesn't matter. Lets just calculate the wordsize even if it
is not needed (it is trivial, anyway).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>