generator: add inverse operation
To better debug the generator itself (..and mostly because it was fun to
implement), this patch adds the inverse-operation to the generator. It
takes a gvariant type and returns the corresponding seed.

The test ./test-generator was updated to always invert its arguments. If
you want to select the operation explicitly, use 'fold/unfold':

    $ ./test-generator fold "a(susso)"
    2665820338153123996776290724
    $ ./test-generator unfold 0002665820338153123996776290724
    a(susso)

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>