writer: advance v_front when finishing a container
Whenever we finish a container, we must merge the front iterator into the
parent container. So far we properly advanced the offset and i_front, but
we must also do so for v_front.

Additionally, we cannot advance i_front as is, but must take the verbatim
value of the child container. Otherwise, we might force folding faults on
the next access.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>