writer: clip front when sealing
Whenever we seal a variant, we must make sure to discard any remaining
space in the front variant. Clip it to make sure the iovecs reflect the
actual data.

Note that the reader worked quite well regardless of this clipping, since
it uses level->offset as root-level input, rather than recalculating the
total iovec size.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>