writer: make c_variant_reserve() take an extra-vec argument
To support c_variant_insert(), we must reserve additional iovecs for
direct use. Extend c_variant_reserve() to actually take an extra-vecs
argument that reserves an additional number of vectors for caller's use.
Pass it through c_variant_append(), so the caller actually gets access to
it.

This also fixes an off-by-one error in the vector-calculation. So far, we
incorrectly assumed that "vec_end - vec_front" gives us the number of
vectors in between, but that's obviously not right.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>