reader: fix size calculation of last non-fixed-sized tuple member
A tuple stores a framing offset for each non-fixed-sized tuple member that
is *NOT* the last member. If the last member is non-fixed-size, its
framing offset is taken to be the size of the container minus the size of
the framing-offset array.

So far, we did adhere to this, but forgot to subtract the framing-offset
array from the overall size. Fix this!

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>