writer: re-calculate vec_front after iovec reallocation
When re-allocating the iovec array, we must re-calculate the vec_tail
pointer as well as the vec_front pointer, as the base might have moved.
So far we only recalculated the tail pointer as its offset might have
changed, but we never accounted for a move of the base pointer.

Fix this and recalculate both pointers every time we reallocate the iovec
array.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>