writer: fix va_arg(args, double) usage
Even though a "double" is the same size as a uint64_t, we must not use
va_arg(args, uint64_t) for it. Calling conventions are free to pass
floating point numbers via floating point registers, so we must not place
any assumptions on it.

Fix the writer to use va_arg(args, double) for GVariant type 'd'.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>