varg: fix array iterator for compound elements
In case of compound array elements like a(), we must make sure we do not
advance the inner array level during iteration. Otherwise, it will have a
different value on each iteration.

This fixes the varg helpers to properly deal with arrays of bigger sizes
than "1", for compound types. For non-compound types this was not an
issue, as we already prevent the iterator modification in
c_variant_varg_next().

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>