tests: add glib-comparison
This adds ./test-glib which runs elaborate marshaling comparisons against
the GVariant implementation in glib. So far, it includes a marshaler for
both glib and cvariant, and compares their binary results. By default, it
runs that for the first 8192 types retrieved via the type generator.

Also included is a print-helper in case something goes wrong.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>